+++ src/rex.sms/NEWS.rst
@@ -5,6 +5,14 @@ REX.SMS Change Log
 .. contents:: Table of Contents
 
 
+2.0.0 (2016-xx-xx)
+==================
+
+* Changed the ``send_sms`` function to require the telephone number of the
+  sender.
+* Added a ShortCodeVal validator for working with Short Codes.
+
+
 1.0.0 (2015-09-30)
 ==================